requirements.txt
# Django is needed to run Django
Django

# gunicorn is needed by Heroku to launch the web server
gunicorn

# Needed for envioronment variables
python-dotenv

# Packages needed for Google Auth
social-auth-app-django
requests

# Package for PostgreSQL
psycopg2

# Packages needed for AWS S3
boto3
django-storages[boto3]